pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyagent"
version = "0.4.0"
description = "The Pandas of AI Agents - Build AI applications in 3 lines or less"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "PyAgent Contributors" }
]
maintainers = [
    { name = "PyAgent Team", email = "pyagent@microsoft.com" }
]
keywords = [
    "ai",
    "agents",
    "llm",
    "rag",
    "artificial-intelligence",
    "machine-learning",
    "openai",
    "anthropic",
    "azure",
    "chatbot",
    "automation",
    "pandas",
    "simple",
    "easy",
    "multi-agent",
    "orchestration",
    "mcp",
    "semantic-kernel",
    "langchain",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
azure = [
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "azure-search-documents>=11.4.0",
]
web = [
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
]
docs = [
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
]
langchain = [
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
]
semantic-kernel = [
    "semantic-kernel>=0.5.0",
]
vector = [
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.0",
    "pinecone-client>=3.0.0",
    "qdrant-client>=1.7.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "azure-identity>=1.15.0",
    "azure-search-documents>=11.4.0",
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "semantic-kernel>=0.5.0",
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/gitpavleenbali/pyagent"
Documentation = "https://github.com/gitpavleenbali/pyagent#readme"
Repository = "https://github.com/gitpavleenbali/pyagent"
Issues = "https://github.com/gitpavleenbali/pyagent/issues"
Changelog = "https://github.com/gitpavleenbali/pyagent/blob/main/docs/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pyagent"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
